image-merge/pyproject.toml
[project]
name = "image-merge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pillow>=10.0.0",
]
